pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-genai>=1.38.0",
    "openai>=1.108.1",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "werkzeug>=3.1.3",
]
